Safety & UART color overflow bug fix

+++ arch/x86/idt.h
@@ -7,4 +7,4 @@
 #define IDT_TRAP_GATE 0x0F
 
 void IDT_init();
-void IDT_set_gate(int interrupt, void* offset, uint16_t segment, uint8_t ist, uint8_t type, uint8_t dpl, bool p);
+void IDT_set_gate(uint8_t interrupt, void* offset, uint16_t segment, uint8_t ist, uint8_t type, uint8_t dpl, bool p);
